Created common base class for DomainObject classes to reduce code duplication.
