+++ tests/src/main/java/com/hazelcast/simulator/tests/map/domain/DomainObjectFactory.java
@@ -5,6 +5,7 @@ import com.hazelcast.simulator.tests.map.SerializationStrategyTest.Strategy;
 public final class DomainObjectFactory {
 
     private final Strategy strategy;
+
     private DomainObjectFactory(Strategy strategy) {
         this.strategy = strategy;
     }
