+++ tests/src/main/java/com/hazelcast/simulator/tests/map/domain/IdentifiedDataSerializableObjectFactory.java
@@ -4,6 +4,7 @@ import com.hazelcast.nio.serialization.DataSerializableFactory;
 import com.hazelcast.nio.serialization.IdentifiedDataSerializable;
 
 public class IdentifiedDataSerializableObjectFactory implements DataSerializableFactory {
+
     public static final int FACTORY_ID = 4000;
 
     @Override
