+++ tests/src/main/java/com/hazelcast/simulator/tests/map/domain/PortableDomainObject.java
@@ -6,74 +6,19 @@ import com.hazelcast.nio.serialization.PortableWriter;
 
 import java.io.IOException;
 
-public class PortableDomainObject implements Portable, DomainObject {
+public class PortableDomainObject extends AbstractDomainObject implements Portable {
+
     public static final int CLASS_ID = 1;
     public static final int FACTORY_ID = PortableObjectFactory.FACTORY_ID;
 
-    private String key;
-    private String stringVal;
-    private double doubleVal;
-    private long longVal;
-    private int intVal;
-
     @Override
-    public String getKey() {
-        return key;
-    }
-
-    @Override
-    public void setKey(String key) {
-        this.key = key;
-    }
-
-    @Override
-    public String getStringVal() {
-        return stringVal;
-    }
-
-    @Override
-    public void setStringVal(String stringVal) {
-        this.stringVal = stringVal;
-    }
-
-    @Override
-    public double getDoubleVal() {
-        return doubleVal;
-    }
-
-    @Override
-    public void setDoubleVal(double doubleVal) {
-        this.doubleVal = doubleVal;
-    }
-
-    @Override
-    public long getLongVal() {
-        return longVal;
-    }
-
-    @Override
-    public void setLongVal(long longVal) {
-        this.longVal = longVal;
-    }
-
-    @Override
-    public int getIntVal() {
-        return intVal;
-    }
-
-    @Override
-    public void setIntVal(int intVal) {
-        this.intVal = intVal;
+    public int getClassId() {
+        return CLASS_ID;
     }
 
     @Override
     public int getFactoryId() {
         return FACTORY_ID;
-    }
-
-    @Override
-    public int getClassId() {
-        return CLASS_ID;
     }
 
     @Override
@@ -93,37 +38,4 @@ public class PortableDomainObject implements Portable, DomainObject {
         longVal = in.readLong("longVal");
         intVal = in.readInt("intVal");
     }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-
-        PortableDomainObject that = (PortableDomainObject) o;
-        if (key != null ? !key.equals(that.key) : that.key != null) {
-            return false;
-        }
-        return true;
-    }
-
-    @Override
-    public int hashCode() {
-        return (key != null ? key.hashCode() : 0);
-    }
-
-    @Override
-    public String toString() {
-        return "HzData {"
-                + "intVal=" + intVal
-                + ", longVal=" + longVal
-                + ", doubleVal=" + doubleVal
-                + ", stringVal='" + stringVal + '\''
-                + ", key='" + key + '\''
-                + '}';
-    }
-
 }
