+++ tests/src/main/java/com/hazelcast/simulator/tests/map/domain/PortableObjectFactory.java
@@ -4,6 +4,7 @@ import com.hazelcast.nio.serialization.Portable;
 import com.hazelcast.nio.serialization.PortableFactory;
 
 public class PortableObjectFactory implements PortableFactory {
+
     public static final int FACTORY_ID = 10000001;
 
     @Override
