+++ tests/src/main/java/com/hazelcast/simulator/tests/map/domain/SerializableDomainObject.java
@@ -2,82 +2,5 @@ package com.hazelcast.simulator.tests.map.domain;
 
 import java.io.Serializable;
 
-public class SerializableDomainObject implements DomainObject, Serializable {
-    private String key;
-    private String stringVal;
-    private double doubleVal;
-    private long longVal;
-    private int intVal;
-
-    @Override
-    public String getKey() {
-        return key;
-    }
-
-    @Override
-    public void setKey(String key) {
-        this.key = key;
-    }
-
-    @Override
-    public String getStringVal() {
-        return stringVal;
-    }
-
-    @Override
-    public void setStringVal(String stringVal) {
-        this.stringVal = stringVal;
-    }
-
-    @Override
-    public double getDoubleVal() {
-        return doubleVal;
-    }
-
-    @Override
-    public void setDoubleVal(double doubleVal) {
-        this.doubleVal = doubleVal;
-    }
-
-    @Override
-    public long getLongVal() {
-        return longVal;
-    }
-
-    @Override
-    public void setLongVal(long longVal) {
-        this.longVal = longVal;
-    }
-
-    @Override
-    public int getIntVal() {
-        return intVal;
-    }
-
-    @Override
-    public void setIntVal(int intVal) {
-        this.intVal = intVal;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-
-        SerializableDomainObject that = (SerializableDomainObject) o;
-        if (key != null ? !key.equals(that.key) : that.key != null) {
-            return false;
-        }
-        return true;
-    }
-
-    @Override
-    public int hashCode() {
-        return (key != null ? key.hashCode() : 0);
-    }
+public class SerializableDomainObject extends AbstractDomainObject implements Serializable {
 }
